Added folder structure and conversion of structural and functional images.
Decided to create bash file with as many options as feasible/possible at start of script.
Complete automization is out of my scope.
For now, we create BIDS structure of folders and already convert structural and functional images.
Possible to add some extra parameters as well as extra scanning sessions.